add repo rev to predicate evaluation
Previously, if a predicate evaluated to a match with a specific repo
rev, the rev would not be added. In practice, I don't think this could
ever be hit because, if a repo rev was specified, the rev filter would
apply after the predicate was evaluated as well, returning only results
that matched the predicate-evaluted repo and the original
revision-narrowed repo.

That said, there may be cases in the future where predicates can
evaluate to specific revisions of a repo, and fixing this now will save
some headaches down the road.