remove CodeExcerptUnhighlighted
CodeExcerptUnhighlighted was the only consumer of the line preview, and
is only used if a dev flag is enabled that disables highlighting. When I
attempted to use this, it seems pretty out of date and appears pretty
broken, so I take that to mean it isn't used. Additionally, if it's a
valuable feature, we can just return the unhighlighted content from the
API rather than the highlighted content, removing the need for a
separate component.