[WIP] internal/gitserver: Get repo metadata from server
I am now stuck here. The intention was to write the metadata received
from the server. But to do that we need access to the ReposDir to
geneate the GitDir of the repo on the client side. Our current design
doesn't allow the client to do this as they are the properties
available on the Server.

We may need to take a step back here and evaluate our next steps.